fix(dcellar-web-ui): disable selection of SPs with invalid status in … (#365)
* fix(dcellar-web-ui): disable selection of SPs with invalid status in SPSelector

* fix(dcellar-web-ui): the sort rule of sps

* fix(dcellar-web-ui): unavailable sp not placed last, dev startup port auto-increment issue

* docs(dcellar-web-ui): update changelog.md
