+++ apps/dcellar-web-ui/package.json
@@ -1,9 +1,9 @@
 {
   "name": "dcellar-web-ui",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "private": false,
   "scripts": {
-    "dev": "next dev -p 3101",
+    "dev": "node ./scripts/dev.js -p 3200",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -56,7 +56,8 @@
     "react-dnd-html5-backend": "16.0.1",
     "viem": "~1.19.11",
     "@wagmi/core": "~1.4.10",
-    "set-interval-async": "~3.0.3"
+    "set-interval-async": "~3.0.3",
+    "radash": "~12.1.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.4.3",
